requirements.txt
typer
qdrant-client
sentence-transformers
python-dotenv
pillow
pymupdf
litellm
tenacity
anthropic